New task on seminar1

+++ seminar1/task2.cpp
@@ -1,6 +1,5 @@
 /* Пример программы печатающей значения PPID и PID для текущего процесса */
 
-#include <sys/types.h>
 #include <unistd.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -12,7 +11,7 @@ int main()
     pid = getpid();
     ppid = getppid();
  
-    printf("My pid = %d, my ppid = %d\n", pid, ppid);
+    printf("My pid = %d, my ppid = %d\n", (int)pid, (int)ppid);
 
     return 0;
 }
